# pkg/fqdn: use LRU in FQDN policy calculation